requirements.txt
pyperclip
pytesseract
Pillow
requests
